+++ backend/foreman/runner.py
@@ -1,14 +1,31 @@
-"""Foreman AI runner: conversation management and main loop."""
+"""Foreman AI runner: conversation management and main loop (embedded)."""
 
 import asyncio
 import json
 import logging
-import os
 from datetime import UTC, datetime
 
 from auth_deps import get_guild_pk
 from database import get_db
 from events import broadcast
+from foreman_core.constants import (
+    MAX_FOREMAN_ROUNDS,
+    MAX_HISTORY_MESSAGES,
+    MAX_TOOL_RESULT_CHARS,
+    _24H_SECS,
+    _HUMAN_TURN_WINDOW,
+    _TERMINAL_STATES,
+)
+from foreman_core.llm import FOREMAN_MODEL, HAS_ANTHROPIC, make_anthropic_client
+from foreman_core.message_utils import (
+    _inject_state_preamble,
+    _serialize_content,
+    _stamp_message_cache_breakpoint,
+    _summarize_task,
+    prune_history,
+    strip_orphaned_tool_results,
+    truncate_tool_result,
+)
 from models import Agent, ForemanTurn, Guild, GuildMember, Message, Task, Worker, live_tasks_filter
 from sqlalchemy import delete, func, select
 from util.tasks import spawn
@@ -16,32 +33,7 @@ from util.tasks import spawn
 from foreman.prompt import build_state_preamble, build_system_blocks, build_system_prompt
 from foreman.tools import FOREMAN_TOOLS, exec_tools
 
-try:
-    import anthropic as _anthropic
-
-    HAS_ANTHROPIC = True
-except ImportError:
-    HAS_ANTHROPIC = False
-
 logger = logging.getLogger(__name__)
-
-FOREMAN_MODEL = os.environ.get("FOREMAN_MODEL", "claude-sonnet-4-6")
-
-# Set FOREMAN_PROVIDER=bedrock to use Amazon Bedrock instead of the Anthropic API.
-# Requires: pip install "anthropic[bedrock]"  +  AWS credentials in env / IAM role.
-# On Bedrock, model IDs use the "anthropic." prefix, e.g.:
-#   anthropic.claude-sonnet-4-5   (Sonnet 4.x on Bedrock)
-#   anthropic.claude-opus-4-5     (Opus 4.x on Bedrock)
-# Check your Bedrock console for exact IDs available in your region.
-FOREMAN_PROVIDER = os.environ.get("FOREMAN_PROVIDER", "anthropic").lower()
-_BEDROCK_REGION = os.environ.get("AWS_DEFAULT_REGION", "us-east-1")
-
-MAX_TOOL_RESULT_CHARS = 8_000  # ~2 k tokens; cap per-result content before storing/sending
-MAX_HISTORY_MESSAGES = 20  # sliding window cap on messages sent to Anthropic
-MAX_FOREMAN_ROUNDS = 10  # safety cap on tool-call rounds per invocation
-_HUMAN_TURN_WINDOW = 5  # how many non-tool-response user turns to load from DB
-_TERMINAL_STATES = frozenset({"done", "failed", "cancelled"})
-_24H_SECS = 86_400
 
 POLL_MIN_SECS = 60  # initial poll interval: 1 minute
 POLL_MAX_SECS = 3600  # maximum poll interval: 60 minutes
@@ -52,217 +44,14 @@ _poll_tasks: dict[str, "asyncio.Task[None]"] = {}
 # Module-level client — reused across calls so the underlying httpx connection
 # pool isn't thrown away every invocation. Lazily initialised so import works
 # without an API key (Anthropic) or AWS credentials (Bedrock).
-_anthropic_client: "_anthropic.AsyncAnthropic | _anthropic.AsyncAnthropicBedrock | None" = None
+_anthropic_client = None
 
 
-def _get_anthropic_client() -> "_anthropic.AsyncAnthropic | _anthropic.AsyncAnthropicBedrock":
+def _get_anthropic_client():
     global _anthropic_client
     if _anthropic_client is None:
-        if FOREMAN_PROVIDER == "bedrock":
-            _anthropic_client = _anthropic.AsyncAnthropicBedrock(
-                aws_region=_BEDROCK_REGION,
-            )
-            logger.info(
-                "Foreman using Amazon Bedrock (region=%s, model=%s)", _BEDROCK_REGION, FOREMAN_MODEL
-            )
-        else:
-            _anthropic_client = _anthropic.AsyncAnthropic()
+        _anthropic_client = make_anthropic_client()
     return _anthropic_client
-
-
-def _inject_state_preamble(messages: list[dict], state_preamble: str) -> None:
-    """Prepend the live-state block to the last user turn (in place).
-
-    Called after history is loaded so the current human turn carries the latest
-    workers/tasks snapshot without persisting that snapshot to the DB.
-    """
-    if not messages or messages[-1]["role"] != "user":
-        return
-    last = messages[-1]
-    content = last["content"]
-    state_block = {"type": "text", "text": state_preamble}
-    if isinstance(content, str):
-        last["content"] = [state_block, {"type": "text", "text": content}]
-    elif isinstance(content, list):
-        last["content"] = [state_block, *content]
-
-
-def _stamp_message_cache_breakpoint(messages: list[dict]) -> None:
-    """Move the messages-level cache breakpoint to the last block of the last turn.
-
-    Clears any prior message-level cache_control first so we never accumulate
-    past the 4-breakpoint API cap (1 used by the system block). Earlier cached
-    prefixes remain readable via the API's 20-block lookback.
-    """
-    for msg in messages:
-        content = msg.get("content")
-        if isinstance(content, list):
-            for block in content:
-                if isinstance(block, dict):
-                    block.pop("cache_control", None)
-    if not messages:
-        return
-    last = messages[-1]
-    content = last.get("content")
-    if isinstance(content, str):
-        last["content"] = [
-            {"type": "text", "text": content, "cache_control": {"type": "ephemeral"}}
-        ]
-        return
-    if isinstance(content, list):
-        for block in reversed(content):
-            if isinstance(block, dict):
-                block["cache_control"] = {"type": "ephemeral"}
-                return
-
-
-def truncate_tool_result(content: str, max_chars: int = MAX_TOOL_RESULT_CHARS) -> str:
-    """Cap a tool result string; append a truncation notice when trimmed."""
-    if len(content) <= max_chars:
-        return content
-    return content[:max_chars] + f"\n… [truncated: {len(content) - max_chars} chars omitted]"
-
-
-def prune_history(messages: list, max_messages: int = MAX_HISTORY_MESSAGES) -> list:
-    """Return at most *max_messages* tail entries, always starting with a user turn."""
-    if len(messages) > max_messages:
-        messages = messages[-max_messages:]
-    while messages and messages[0]["role"] != "user":
-        messages = messages[1:]
-    return messages
-
-
-def strip_orphaned_tool_results(messages: list[dict]) -> list[dict]:
-    """Remove any tool_result blocks (and their containing user message if it becomes empty)
-    whose tool_use_id has no matching tool_use block in the immediately-preceding
-    assistant message. Also remove the dangling tool_use blocks from the assistant
-    message that have no corresponding tool_result.
-    Returns a cleaned copy of the messages list.
-    """
-    import copy
-
-    out = copy.deepcopy(messages)
-    i = 0
-    while i < len(out):
-        msg = out[i]
-        content = msg.get("content")
-
-        if msg["role"] == "assistant" and isinstance(content, list):
-            # Collect tool_result IDs from the immediately-following user message.
-            if i + 1 < len(out):
-                nxt = out[i + 1]
-                if nxt["role"] == "user" and isinstance(nxt.get("content"), list):
-                    result_ids = {
-                        b["tool_use_id"]
-                        for b in nxt["content"]
-                        if isinstance(b, dict) and b.get("type") == "tool_result"
-                    }
-                else:
-                    result_ids = set()
-            else:
-                result_ids = set()
-
-            new_content = [
-                b
-                for b in content
-                if not (isinstance(b, dict) and b.get("type") == "tool_use")
-                or b.get("id") in result_ids
-            ]
-            if new_content != content:
-                if not new_content:
-                    out.pop(i)
-                    continue
-                msg["content"] = new_content
-
-        elif msg["role"] == "user" and isinstance(content, list):
-            # Collect tool_use IDs from the immediately-preceding assistant message.
-            if i > 0:
-                prev = out[i - 1]
-                if prev["role"] == "assistant" and isinstance(prev.get("content"), list):
-                    valid_ids = {
-                        b["id"]
-                        for b in prev["content"]
-                        if isinstance(b, dict) and b.get("type") == "tool_use"
-                    }
-                else:
-                    valid_ids = set()
-            else:
-                valid_ids = set()
-
-            new_content = [
-                b
-                for b in content
-                if not (isinstance(b, dict) and b.get("type") == "tool_result")
-                or b.get("tool_use_id") in valid_ids
-            ]
-            if new_content != content:
-                if not new_content:
-                    out.pop(i)
-                    continue
-                msg["content"] = new_content
-
-        i += 1
-
-    # Re-enforce starts-with-user and handle cascading orphans at the head.
-    # Removing a leading orphaned user message exposes a leading assistant; removing
-    # that assistant may expose another user message whose tool_results are now
-    # orphaned (their matching assistant was just dropped). Loop until stable.
-    while out:
-        if out[0]["role"] != "user":
-            out.pop(0)
-            continue
-        content = out[0].get("content")
-        if isinstance(content, list):
-            new_content = [
-                b for b in content if not (isinstance(b, dict) and b.get("type") == "tool_result")
-            ]
-            if not new_content:
-                out.pop(0)
-                continue
-            if new_content != content:
-                out[0]["content"] = new_content
-        break
-    return out
-
-
-def _summarize_task(task: dict, cutoff_ts: float) -> dict | None:
-    """Return a (possibly stripped) task dict, or None to exclude it.
-
-    Terminal tasks older than 24 h are dropped; terminal tasks within 24 h lose
-    their ``description`` field to keep context lean.  Non-terminal tasks are
-    returned unchanged.
-    """
-    state = task.get("state", "")
-    if state not in _TERMINAL_STATES:
-        return task
-    finished_at = task.get("finished_at")
-    if finished_at:
-        try:
-            finished_ts = datetime.fromisoformat(finished_at.replace("Z", "+00:00")).timestamp()
-            if finished_ts < cutoff_ts:
-                return None  # older than 24 h — drop entirely
-        except (ValueError, AttributeError):
-            pass
-    # Within 24 h or undatable: compact summary without description
-    return {k: v for k, v in task.items() if k != "description"}
-
-
-def _serialize_content(content) -> str:
-    """Convert SDK content objects or dicts to a JSON string for DB storage."""
-    if isinstance(content, str):
-        return json.dumps(content)
-    if isinstance(content, list):
-        blocks = []
-        for b in content:
-            if isinstance(b, dict):
-                blocks.append(b)
-            else:
-                try:
-                    blocks.append(b.model_dump())
-                except AttributeError:
-                    blocks.append({"type": str(getattr(b, "type", "unknown")), "raw": str(b)})
-        return json.dumps(blocks)
-    return json.dumps(str(content))
 
 
 async def _get_guild_user_id(guild_id: str) -> str | None:
