+++ foreman_core/foreman_core/__init__.py
@@ -1,0 +1,1 @@
+"""Shared core library for Pioneer Square foreman (embedded + standalone)."""
