+++ foreman_core/pyproject.toml
@@ -1,0 +1,17 @@
+[build-system]
+requires = ["setuptools>=68", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "foreman-core"
+version = "0.1.0"
+description = "Shared core library for Pioneer Square foreman (embedded + standalone)."
+requires-python = ">=3.11"
+dependencies = []
+
+[project.optional-dependencies]
+anthropic = ["anthropic>=0.49"]
+bedrock = ["anthropic[bedrock]>=0.49"]
+
+[tool.setuptools.packages.find]
+include = ["foreman_core*"]
